Committing fix to both the met-6.0 bugfix branch and the development trunk.  These changes are needed for the Air Force 2017 tier 2 evaluation.  This allows for non-integer level values such as soil temperature 0.1 meters below ground versus 1.0 meters below ground.  I ported over logic from the wgrib2 to apply scale factors to level values.
Former-commit-id: 0437e4dae183bcd56a1b43d397dd5364af1dd92e

+++ src/libcode/vx_data2d/var_info.cc
@@ -428,8 +428,8 @@ void VarInfo::set_level_info_grib(Dictionary & dict){
    //  otherwise, just the numeric value
    ConcatString lvl_name;
    if( lt == LevelType_Accum )   lvl_name.format("%s%s",    lvl_type.data(), lvl_val1.data());
-   else if( !is_eq(lvl2, -1.0) ) lvl_name.format("%s%d-%d", lvl_type.data(), (int)lvl2, (int)lvl1);
-   else                          lvl_name.format("%s%d",    lvl_type.data(), (int)lvl1);
+   else if( !is_eq(lvl2, -1.0) ) lvl_name.format("%s%g-%g", lvl_type.data(), lvl2, lvl1);
+   else                          lvl_name.format("%s%g",    lvl_type.data(), lvl1);
 
    //  set the level information
    Level.set_type(lt);
